Adjusted index.html, changed main.js to align with homework examples, and debugged main_with_debug.js.
In index.html: changed header color to "bisque", changed title to "Greetings!", streamlined the code that adds "Hello World!" at the end of the title, and changed  which JavaScript file the HTML reads (now it's main_with_debug.js).

In main.js: changed most of the code to align with the homework reading examples.

In main_with_debug.js: debugged the code so it works.